Rename specs_batteries module to ecs
