+++ src/asset_manager/asset_manager.rs
@@ -24,7 +24,7 @@ use std::str;
 use std::sync::RwLockReadGuard;
 
 // self imports
-use specs_batteries::components::rendering::{Mesh, Renderable, Texture, TextureLoadData};
+use ecs::components::rendering::{Mesh, Renderable, Texture, TextureLoadData};
 use renderer::VertexPosNormal;
 
 // external imports
