+++ src/ecs/components/rendering/mod.rs
