+++ src/ecs/components/rendering/texture.rs
